feat : 명예의 전당  userId → nickname 매핑
