+++ app/(pola)/(junior)/hall-of-fame/Ranking.module.css
@@ -1,15 +1,16 @@
 /* 랭킹 1,2,3 */
 .rankingWrap {
   background: #2c2c2c;
-  padding: 0rem 2rem 2rem;
   width: 100%;
+  height: 100%;
   /* border-radius: 24px;padding: 32px 24px;box-shadow: 0 4px 24px #0003; */
 }
-
 .rankingTopWrap {
+  padding-top: 3rem;
   position: relative;
   overflow: hidden;
-  min-height: 13rem;
+  min-height: 16rem;
+  box-shadow: 1px -10px 12px 3px #f0f0f0;
 }
 
 .rankingTopWrap ul {
@@ -52,13 +53,35 @@
   content: '';
   position: absolute;
   top: 10px;
-  width: 30px;
-  height: 30px;
-  background-color: #fff;
+  background-size: contain;
+  background-repeat: no-repeat;
+  background-position: center;
   animation: rotateY360 2s linear infinite;
   transform-style: preserve-3d;
   backface-visibility: hidden;
 }
+
+.rankingTopWrap ul li:nth-child(1):before {
+  background-image: url('/images/ranking/num_1.png');
+  width: 4.5rem;
+  height: 4.5rem;
+}
+
+.rankingTopWrap ul li:nth-child(2):before {
+  background-image: url('/images/ranking/num_2.png');
+  width: 3.5rem;
+  height: 3.5rem;
+  top: 19px;
+}
+
+.rankingTopWrap ul li:nth-child(3):before {
+  background-image: url('/images/ranking/num_3.png');
+  width: 2.5rem;
+  height: 2.5rem;
+  top: 25px;
+}
+
+
 
 .rankingTopWrap ul li .rankingTopName{
   display: inline-block;
@@ -72,12 +95,13 @@
 .rankingTopWrap ul li:nth-child(1) {
   font-size: 1.3em;
   font-weight: bold;
-  transform: translateY(0);
+  transform: translateY(-5px)scale(1.1);
   order: 1;
 }
 
 .rankingTopWrap ul li:nth-child(3) {
   order: 3;
+  transform: translateY(25px)scale(.9);
 }
 
 @keyframes rankingTopWrap {
@@ -100,6 +124,20 @@
   list-style: none;
   margin: 0;
   padding: 0;
+  opacity: 0;
+  animation: rankingBtmWrap 1s 1s ease-in-out forwards;
+  transition: all .2s;
+}
+
+@keyframes rankingBtmWrap {
+  from {
+    opacity: 0;
+    transform: translateY(100%);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
 }
 
 .rankingBtmWrap .rankingBtmItem {
@@ -107,7 +145,7 @@
   justify-content: space-between;
   color: #fff;
   list-style: none;
-  padding:2rem;
+  padding:1.2rem;
   align-items: center;
 }
 
@@ -117,11 +155,15 @@
   height: 40px;
   border-radius: 50%;
   object-fit: cover;
-  margin-right: 1rem;
+  margin-right: .5rem;
   border: 2px solid #fff;
   background: #222;
 }
 
+.rankingBtmRank{
+  display: inline-block;
+  margin-right: .5rem;
+}
 /* 하단 랭킹 이미지와 등수 컨테이너 */
 .rankingBtmImgContainer {
   display: flex;
@@ -179,8 +221,8 @@
 /* 조명 */
 .circleLights {
   background: #fff;
-  width: 200px;
-  height: 200px;
+  width: 300px;
+  height: 300px;
   position: absolute;
   top: 50%;
   left: 50%;
@@ -214,3 +256,16 @@
     opacity: 0;
   }
 } 
+
+/* 반응형 쿼리 - 최소 370px */
+@media (max-width: 370px) {
+  .rankingTopWrap  {margin: 0 -3rem;}
+  .rankingWrap{  padding:0rem; min-height: 8rem;}
+  .rankingWrap > * {transform: scale(0.8);}
+
+  .rankingBtmWrap{margin: -5rem -3rem 0rem;}
+}
+
+.rankingTopScore {
+  margin-top: 0.4rem;
+}
