--- app/(pola)/main/layout.tsx
@@ -1,7 +1,0 @@
-export default function MainLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <>{children}</>;
-}
